+++ DelvUI/Config/ConfigurationManager.cs
@@ -465,6 +465,7 @@ namespace DelvUI.Config
             typeof(MPTickerConfig),
             typeof(GridConfig),
             typeof(PullTimerConfig),
+            typeof(LimitBreakConfig),
 
             typeof(ImportConfig)
         };